Various improvments for 64-bit C and 32-bit version.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@209 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/basetsd.h
@@ -106,10 +106,10 @@ extern "C" {
 #define LongToPtr(l) ((VOID *)(LONG_PTR)((long)l))
 #define ULongToPtr(ul) ((VOID *)(ULONG_PTR)((unsigned long)ul))
 
-  static __inline void *PtrToPtr64(const void *p) { return((void *) (unsigned __int64) (ULONG_PTR)p); }
-  static __inline void *Ptr64ToPtr(const void *p) { return((void *) (ULONG_PTR) (unsigned __int64) p); }
+  static __inline void *PtrToPtr64(const void *p) { return((void *) (ULONG_PTR)p); }
+  static __inline void *Ptr64ToPtr(const void *p) { return((void *) (ULONG_PTR) p); }
   static __inline void *HandleToHandle64(const void *h) { return((void *) h); }
-  static __inline void *Handle64ToHandle(const void *h) { return((void *) (ULONG_PTR) (unsigned __int64) h); }
+  static __inline void *Handle64ToHandle(const void *h) { return((void *) (ULONG_PTR) h); }
 
 #define Ptr32ToPtr(p) ((void *) p)
 #define Handle32ToHandle(h) (Ptr32ToPtr(h))
